Refactor opcode labels to late-compute positions
Since we will be rewritting the bytecode block, this ends up moving
around targets of jumps and exception protected sections (ie. we need to
compute these positions as late as possible.)

This patch tries to compute the positions only when they need to be
written to the image. It also splits Label into Label and Ref, where
label computes offsets given a base (for jumps), and Refs are used for
identifying exception protected sections.