wip vararg: fixed buggy fp layout.
The calculations for fp in setup_fp() were not correct when ss==0 and
argc==0 due to the choice of fp position (pointing to the last argument
or topmost cell in local storage). When storage size and argc are 0, FP
points to local data in the previous call frame. Then, setup_fp() copies
RP/DP correctly to new allocated env, however, it calculates the new FP
to be one address before the allocated env -- pointing to an invalid
address and effectivelly making the region inacessible and subject to GC
collect -- which caused pernicious bugs.

Finally, tests are working.